refac collaborator seed into service to use inside rake task
